Split tests to several files
